Moved fiber photometry to a dedicated section.
