+++ docs/api/interfaces.fiber_photometry.rst
@@ -1,0 +1,14 @@
+Fiber Photometry
+================
+
+Base Fiber Photometry
+---------------------
+.. automodule:: neuroconv.datainterfaces.fiber_photometry.basefiberphotometryinterface
+
+Doric Fiber Photometry
+----------------------
+.. automodule:: neuroconv.datainterfaces.fiber_photometry.doric.doricfiberphotometrydatainterface
+
+TDT Fiber Photometry
+--------------------
+.. automodule:: neuroconv.datainterfaces.fiber_photometry.tdt.tdtfiberphotometrydatainterface
