+++ docs/api/interfaces.rst
@@ -7,6 +7,7 @@ Interfaces
     interfaces.ecephys
     interfaces.icephys
     interfaces.ophys
+    interfaces.fiber_photometry
     interfaces.behavior
     interfaces.events
     interfaces.text
