+++ src/neuroconv/datainterfaces/fiber_photometry/tdt/_tdt_mixin.py
