+++ tests/test_on_data/fiber_photometry/fiber_photometry_metadata.yaml
