+++ tests/test_on_data/fiber_photometry/test_fiber_photometry_interfaces.py
@@ -13,7 +13,7 @@ from neuroconv.datainterfaces import (
     DoricFiberPhotometryInterface,
     TDTFiberPhotometryInterface,
 )
-from neuroconv.datainterfaces.ophys.tdt_fp.tdtfiberphotometrydatainterface import (
+from neuroconv.datainterfaces.fiber_photometry.tdt.tdtfiberphotometrydatainterface import (
     _TDTFiberPhotometryInterfaceMultiSeries,
 )
 from neuroconv.tools.testing.data_interface_mixins import (
@@ -713,106 +713,104 @@ _DORIC_STREAM_NAMES = sorted(
 )
 
 _DORIC_METADATA = {
-    "Ophys": {
-        "FiberPhotometry": {
-            "OpticalFiberModels": [
-                {
-                    "name": "optical_fiber_model",
-                    "manufacturer": "Doric Lenses",
-                    "numerical_aperture": 0.48,
-                    "core_diameter_in_um": 400.0,
-                }
-            ],
-            "OpticalFibers": [
-                {
-                    "name": "optical_fiber",
-                    "model": "optical_fiber_model",
-                    "fiber_insertion": {"depth_in_mm": 2.8},
-                }
-            ],
-            "ExcitationSourceModels": [
-                {
-                    "name": "excitation_source_model",
-                    "manufacturer": "Doric Lenses",
-                    "source_type": "LED",
-                    "excitation_mode": "one-photon",
-                }
-            ],
-            "ExcitationSources": [
-                {
-                    "name": "excitation_source_465nm",
-                    "model": "excitation_source_model",
-                },
-                {
-                    "name": "excitation_source_405nm",
-                    "model": "excitation_source_model",
-                },
-            ],
-            "PhotodetectorModels": [
-                {
-                    "name": "photodetector_model",
-                    "manufacturer": "Doric Lenses",
-                    "detector_type": "photodiode",
-                }
-            ],
-            "Photodetectors": [
-                {
-                    "name": "photodetector",
-                    "model": "photodetector_model",
-                }
-            ],
-            "FiberPhotometryIndicators": [
-                {
-                    "name": "green_fluorophore",
-                    "description": "GCaMP7b calcium indicator.",
-                    "label": "GCaMP7b",
-                }
-            ],
-            "FiberPhotometryTable": {
-                "name": "fiber_photometry_table",
-                "description": "Fiber photometry acquisition metadata.",
-                "rows": [
-                    {
-                        "name": "0",
-                        "location": "DMS",
-                        "excitation_wavelength_in_nm": 465.0,
-                        "emission_wavelength_in_nm": 525.0,
-                        "indicator": "green_fluorophore",
-                        "optical_fiber": "optical_fiber",
-                        "excitation_source": "excitation_source_465nm",
-                        "photodetector": "photodetector",
-                    },
-                    {
-                        "name": "1",
-                        "location": "DMS",
-                        "excitation_wavelength_in_nm": 405.0,
-                        "emission_wavelength_in_nm": 525.0,
-                        "indicator": "green_fluorophore",
-                        "optical_fiber": "optical_fiber",
-                        "excitation_source": "excitation_source_405nm",
-                        "photodetector": "photodetector",
-                    },
-                ],
+    "FiberPhotometry": {
+        "OpticalFiberModels": [
+            {
+                "name": "optical_fiber_model",
+                "manufacturer": "Doric Lenses",
+                "numerical_aperture": 0.48,
+                "core_diameter_in_um": 400.0,
+            }
+        ],
+        "OpticalFibers": [
+            {
+                "name": "optical_fiber",
+                "model": "optical_fiber_model",
+                "fiber_insertion": {"depth_in_mm": 2.8},
+            }
+        ],
+        "ExcitationSourceModels": [
+            {
+                "name": "excitation_source_model",
+                "manufacturer": "Doric Lenses",
+                "source_type": "LED",
+                "excitation_mode": "one-photon",
+            }
+        ],
+        "ExcitationSources": [
+            {
+                "name": "excitation_source_465nm",
+                "model": "excitation_source_model",
             },
-            "FiberPhotometryResponseSeries": [
+            {
+                "name": "excitation_source_405nm",
+                "model": "excitation_source_model",
+            },
+        ],
+        "PhotodetectorModels": [
+            {
+                "name": "photodetector_model",
+                "manufacturer": "Doric Lenses",
+                "detector_type": "photodiode",
+            }
+        ],
+        "Photodetectors": [
+            {
+                "name": "photodetector",
+                "model": "photodetector_model",
+            }
+        ],
+        "FiberPhotometryIndicators": [
+            {
+                "name": "green_fluorophore",
+                "description": "GCaMP7b calcium indicator.",
+                "label": "GCaMP7b",
+            }
+        ],
+        "FiberPhotometryTable": {
+            "name": "fiber_photometry_table",
+            "description": "Fiber photometry acquisition metadata.",
+            "rows": [
                 {
-                    "name": "signal_exc1_roi1",
-                    "description": "465 nm channel, ROI 1.",
-                    "stream_name": "BBC300_ROISignals_Series0001_CAM1EXC1_ROI01",
-                    "unit": "a.u.",
-                    "fiber_photometry_table_region": [0],
-                    "fiber_photometry_table_region_description": "Row 0.",
+                    "name": "0",
+                    "location": "DMS",
+                    "excitation_wavelength_in_nm": 465.0,
+                    "emission_wavelength_in_nm": 525.0,
+                    "indicator": "green_fluorophore",
+                    "optical_fiber": "optical_fiber",
+                    "excitation_source": "excitation_source_465nm",
+                    "photodetector": "photodetector",
                 },
                 {
-                    "name": "signal_exc2_roi1",
-                    "description": "405 nm channel, ROI 1.",
-                    "stream_name": "BBC300_ROISignals_Series0001_CAM1EXC2_ROI01",
-                    "unit": "a.u.",
-                    "fiber_photometry_table_region": [1],
-                    "fiber_photometry_table_region_description": "Row 1.",
+                    "name": "1",
+                    "location": "DMS",
+                    "excitation_wavelength_in_nm": 405.0,
+                    "emission_wavelength_in_nm": 525.0,
+                    "indicator": "green_fluorophore",
+                    "optical_fiber": "optical_fiber",
+                    "excitation_source": "excitation_source_405nm",
+                    "photodetector": "photodetector",
                 },
             ],
-        }
+        },
+        "FiberPhotometryResponseSeries": [
+            {
+                "name": "signal_exc1_roi1",
+                "description": "465 nm channel, ROI 1.",
+                "stream_name": "BBC300_ROISignals_Series0001_CAM1EXC1_ROI01",
+                "unit": "a.u.",
+                "fiber_photometry_table_region": [0],
+                "fiber_photometry_table_region_description": "Row 0.",
+            },
+            {
+                "name": "signal_exc2_roi1",
+                "description": "405 nm channel, ROI 1.",
+                "stream_name": "BBC300_ROISignals_Series0001_CAM1EXC2_ROI01",
+                "unit": "a.u.",
+                "fiber_photometry_table_region": [1],
+                "fiber_photometry_table_region_description": "Row 1.",
+            },
+        ],
     }
 }
 
@@ -964,7 +962,7 @@ class TestTDTFiberPhotometryInterfaceSingleSeries(FiberPhotometryInterfaceTestMi
     def check_extracted_metadata(self, metadata: dict):
         assert "session_start_time" in metadata["NWBFile"]
         # "Signal" is the metadata_key under which this interface's response-series metadata lives.
-        assert "Signal" in metadata["Ophys"]["FiberPhotometry"]
+        assert "Signal" in metadata["FiberPhotometry"]
 
     def test_get_available_streams(self):
         streams = self.data_interface_cls.get_available_streams(folder_path=self.interface_kwargs["folder_path"])
@@ -979,4 +977,4 @@ class TestTDTFiberPhotometryInterfaceSingleSeries(FiberPhotometryInterfaceTestMi
         # With no explicit metadata_key, it is derived from stream_names (as in ScanImage).
         interface = self.data_interface_cls(folder_path=self.interface_kwargs["folder_path"], stream_names="_405R")
         assert interface.metadata_key == "fiber_photometry_405r"
-        assert interface.metadata_key in interface.get_metadata()["Ophys"]["FiberPhotometry"]
+        assert interface.metadata_key in interface.get_metadata()["FiberPhotometry"]
